refactor(systems-grid): extract pull-to-refresh/pinch gestures into a part file
Move the pointer-handling slice of _SystemCardGridViewState
(_handlePointerDown/Move/Up/Cancel, _triggerRefresh, _adjustGridDensity,
_showCardSizeLabel) verbatim into a new part file,
my_systems_grid/pull_to_refresh.dart, as `extension _PullToRefresh`.

Pure source-split, no behaviour or public-API change. Moved bodies are
byte-identical to main (token-stream verified). All state stays on the host
State; the 1 setState call routes through the host `rebuild()` bridge
(State.setState is @protected). Host 1108 -> 1011; new file 111 lines.

Closes out D controller parts (#152/#153/this). Local gate: dart format
clean, flutter analyze "No issues found", 207/207.
