+++ lib/screens/systems_screen/my_systems_section/my_systems_grid.dart
@@ -42,6 +42,7 @@ import 'system_list_builder.dart';
 
 part 'my_systems_grid/gamepad_grid_nav.dart';
 part 'my_systems_grid/theme_background.dart';
+part 'my_systems_grid/pull_to_refresh.dart';
 
 /// Primary widget for the 'My Systems' view, supporting both Grid and Carousel layouts.
 ///
@@ -860,104 +861,6 @@ class _SystemCardGridViewState extends State<SystemCardGridView> {
         );
       },
     );
-  }
-
-  void _handlePointerDown(PointerDownEvent event) {
-    _activePointers[event.pointer] = event.position;
-  }
-
-  void _handlePointerMove(PointerMoveEvent event) {
-    _activePointers[event.pointer] = event.position;
-    if (_activePointers.length < 2) return;
-
-    final now = DateTime.now();
-    if (_lastPinchTime != null &&
-        now.difference(_lastPinchTime!).inMilliseconds < 120) {
-      return;
-    }
-
-    final positions = _activePointers.values.toList();
-    final distance = (positions[0] - positions[1]).distance;
-
-    if (_lastPinchDistance != null) {
-      final deltaDistance = distance - _lastPinchDistance!;
-      if (deltaDistance > 35) {
-        _adjustGridDensity(1);
-        _lastPinchDistance = distance;
-        _lastPinchTime = now;
-      } else if (deltaDistance < -35) {
-        _adjustGridDensity(-1);
-        _lastPinchDistance = distance;
-        _lastPinchTime = now;
-      }
-    } else {
-      _lastPinchDistance = distance;
-    }
-  }
-
-  void _handlePointerUp(PointerUpEvent event) {
-    _activePointers.remove(event.pointer);
-    if (_activePointers.length < 2) {
-      _lastPinchDistance = null;
-    }
-    if (_activePointers.isEmpty && _pullReady) {
-      _pullReady = false;
-      _pullProgress.value = 0.0;
-      _triggerRefresh();
-    }
-  }
-
-  void _handlePointerCancel(PointerCancelEvent event) {
-    _activePointers.remove(event.pointer);
-    if (_activePointers.length < 2) {
-      _lastPinchDistance = null;
-    }
-    if (_activePointers.isEmpty && _pullReady) {
-      _pullReady = false;
-      _pullProgress.value = 0.0;
-      _triggerRefresh();
-    }
-  }
-
-  /// Triggers ROM directory rescan when pull-to-refresh reaches 100%.
-  void _triggerRefresh() {
-    try {
-      final configProvider = context.read<SqliteConfigProvider>();
-      if (!configProvider.isScanning) {
-        SfxService().playNavSound();
-        configProvider.scanSystems();
-      }
-    } catch (_) {
-      // Ignore if context/provider is no longer valid.
-    }
-  }
-
-  /// Adjusts grid column density based on pinch gesture direction.
-  void _adjustGridDensity(int delta) {
-    try {
-      final provider = context.read<SqliteConfigProvider>();
-      final sizes = ['S', 'M', 'L', 'XL'];
-      final currentIndex = sizes.indexOf(provider.config.systemGridColumns);
-      if (currentIndex == -1) return;
-      final newIndex = (currentIndex + delta).clamp(0, sizes.length - 1);
-      if (newIndex != currentIndex) {
-        final newSize = sizes[newIndex];
-        provider.updateSystemGridColumns(newSize);
-        _cols = Responsive.getSystemsCrossAxisCountFromSize(newSize);
-        _cachedVirtualGrid = null;
-        _cachedGridCols = null;
-        _showCardSizeLabel(newSize);
-        setState(() {});
-      }
-    } catch (_) {}
-  }
-
-  void _showCardSizeLabel(String size) {
-    _cardSizeLabelTimer?.cancel();
-    _cardSizeLabel.value = size;
-    _cardSizeLabelTimer = Timer(const Duration(milliseconds: 1200), () {
-      _cardSizeLabel.value = null;
-    });
   }
 
   /// Renders a non-linear grid by manually positioning components according to their spans.
